fix: move to protocol/contacts
