+++ protocol/message_persistence.go
@@ -14,8 +14,8 @@ import (
 
 	"github.com/status-im/markdown"
 
-	"github.com/status-im/status-go/pkg/contacts"
 	"github.com/status-im/status-go/protocol/common"
+	"github.com/status-im/status-go/protocol/contacts"
 	"github.com/status-im/status-go/protocol/protobuf"
 )
