+++ protocol/messenger_group_chat_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/status-im/status-go/crypto"
 	userimage "github.com/status-im/status-go/images"
 	multiaccountscommon "github.com/status-im/status-go/multiaccounts/common"
-	"github.com/status-im/status-go/pkg/contacts"
+	"github.com/status-im/status-go/protocol/contacts"
 
 	"github.com/status-im/status-go/multiaccounts/settings"
 	"github.com/status-im/status-go/protocol/protobuf"
