+++ protocol/messenger_maps.go
@@ -6,7 +6,7 @@ import (
 	"go.uber.org/zap"
 
 	messagingtypes "github.com/status-im/status-go/messaging/types"
-	"github.com/status-im/status-go/pkg/contacts"
+	"github.com/status-im/status-go/protocol/contacts"
 	"github.com/status-im/status-go/protocol/protobuf"
 )
