+++ protocol/messenger_test.go
@@ -19,9 +19,9 @@ import (
 	"github.com/status-im/status-go/images"
 	"github.com/status-im/status-go/internal/timesource"
 	"github.com/status-im/status-go/multiaccounts/settings"
-	"github.com/status-im/status-go/pkg/contacts"
 	"github.com/status-im/status-go/protocol/common"
 	"github.com/status-im/status-go/protocol/communities"
+	"github.com/status-im/status-go/protocol/contacts"
 	"github.com/status-im/status-go/protocol/protobuf"
 	"github.com/status-im/status-go/protocol/requests"
 	v1protocol "github.com/status-im/status-go/protocol/v1"
